+++ src/components/ImportButton/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button as RAButton } from 'react-admin';
+import {Button as RAButton, useRefresh} from 'react-admin';
 import GetAppIcon from '@material-ui/icons/GetApp';
 import { useDataProvider } from 'react-admin';
 import { processCsvFile } from './csvExtractor';
@@ -35,6 +35,7 @@ export const ImportButton = (props) => {
   const [dialogStatus, setDialogStatus] = React.useState('base');
   const [csvValidationNotifications, setCsvValidationNotifications] = React.useState([]);
   const dataProvider = useDataProvider();
+  const refresh = useRefresh()
 
   const openImportDialog = () => {
     setOpen(true);
@@ -50,17 +51,15 @@ export const ImportButton = (props) => {
 
   const handleClose = () => {
     clear();
-
     setOpen(false);
   };
 
   const handleImportDelimiterChange = (e) => {
     setDelimiter(e.target.value);
-
     clear();
   };
 
-  const handleSubmit = (callback) => {
+  const handleSubmit = (callback, afterSucces) => {
     setImporting(true);
 
     let apiValidationErrors = [];
@@ -80,9 +79,9 @@ export const ImportButton = (props) => {
       })),
     ).then(() => {
       setCsvValidationNotifications(apiValidationErrors);
-
       setImporting(false);
       setDialogStatus('importFinished');
+      refresh();
     });
   };