+++ src/resources/user.jsx
@@ -1,6 +1,18 @@
 // in Users.js
 import React from 'react';
-import { List, Datagrid, Edit, Create, SimpleForm, DateField, TextField, EditButton, TextInput, Filter } from 'react-admin';
+import {
+    List,
+    Datagrid,
+    Edit,
+    Create,
+    SimpleForm,
+    DateField,
+    TextField,
+    EditButton,
+    TextInput,
+    Filter,
+    useNotify, useRefresh, useRedirect
+} from 'react-admin';
 import PersonIcon from '@material-ui/icons/Person';
 //import EditableDatagrid from './react-data-grid/index.jsx';
 
@@ -12,7 +24,7 @@ const UserFilter = (props) => (
 );
 
 export const UserList = (props) => (
-    <List {...props} filters={<UserFilter />}>
+    <List {...props} filters={<UserFilter />} sort={{field: 'id', order: 'DESC'}}>
         <Datagrid>
             <TextField source="id" />
             <TextField source="role" />
@@ -41,15 +53,23 @@ export const UserEdit = (props) => (
     </Edit>
 );
 
-export const UserCreate = (props) => (
-    <Create title="Create a User" {...props}>
+export const UserCreate = (props) => {
+    const notify = useNotify();
+    const refresh = useRefresh();
+    const redirect = useRedirect();
+
+    return <Create title="Create a User" {...props} onSuccess={() => {
+        notify('ra.notification.created', 'info', {smart_count: 1});
+        redirect('/user');
+        refresh()
+    }}>
         <SimpleForm>
-          <TextInput disabled source="id" variant="outlined" fullWidth />
-          <TextInput disabled source="role" variant="outlined" fullWidth  />
-          <TextInput source="firstName" variant="outlined" fullWidth  />
-          <TextInput source="lastName" variant="outlined" fullWidth  />
-          <TextInput source="email" variant="outlined" fullWidth  />
-          <TextInput source="zipCode" variant="outlined" fullWidth  />
+            <TextInput disabled source="id" variant="outlined" fullWidth/>
+            <TextInput disabled source="role" variant="outlined" fullWidth/>
+            <TextInput source="firstName" variant="outlined" fullWidth/>
+            <TextInput source="lastName" variant="outlined" fullWidth/>
+            <TextInput source="email" variant="outlined" fullWidth/>
+            <TextInput source="zipCode" variant="outlined" fullWidth/>
         </SimpleForm>
     </Create>
-);
+};